fix bugs with error catch
