+++ middlewares/isEmptyReq.js
@@ -1,13 +1,9 @@
 import HttpError from "../utils/index.js";
 
-const isEmptyReq = (schema) => {
-  const func = (req, res, next) => {
-    if (Object.keys(req.body).length === 0)
-      throw HttpError(400, "missing fields");
-    next();
-  };
-
-  return func;
+const isEmptyReq = (req, res, next) => {
+  if (Object.keys(req.body).length === 0)
+    next(HttpError(400, "missing fields"));
+  next();
 };
 
 export default isEmptyReq;
